x86/microcode/intel: Add ability to update microcode even if rev is unchanged
This comes in handy for testing without the need for a new microcode file.
Default is OFF. It can be switched dynamically at run time via debugfs file
/sys/kernel/debug/microcode/load_same.

NOT_FOR_INCLUSION:
Leave it to the discretion of Boris if its suitable for inclusion. It will
at least serve to validate some parts of the series without the need for a
new microcode, since it goes through all the renedezvous parts except, but
reloads the same microcode that was loaded early/late.

Add a debugfs to permit allowing a microcode to be loaded even if the
revision hasn't changed.

Reviewed-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Ashok Raj <ashok.raj@intel.com>