x86/ipi: Support sending NMI_VECTOR as self ipi
apic->send_IPI_self() can be used to send any general vector as a self IPI.
The function uses the "Destination Shorthand", but it can't be used to send a
vector with delivery mode as NMI[1].

Add a helper to check if the vector is NMI_VECTOR and call the send_IPI
without using the destination shorthand.

[1] In the SDM Vol 3a, Chapter 10,  Advanced Programmable Interrupt
Controller. It notes that the self "Destination Shorthand" is invalid for
NMIs.

Suggested-by: Ashok Raj <ashok.raj@intel.com>
Signed-off-by: Jacob Pan <jacob.jun.pan@linux.intel.com>
Co-developed-by: Ashok Raj <ashok.raj@intel.com>
Signed-off-by: Ashok Raj <ashok.raj@intel.com>
Reviewed-by: Tony Luck <tony.luck@intel.com>
Cc: LKML <linux-kernel@vger.kernel.org>
Cc: x86 <x86@kernel.org>
Cc: Tony Luck <tony.luck@intel.com>
Cc: Ingo Molnar <mingo@kernel.org>
Cc: Dave Hansen <dave.hansen@intel.com>
Cc: Alison Schofield <alison.schofield@intel.com>
Cc: Reinette Chatre <reinette.chatre@intel.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Stefan Talpalaru <stefantalpalaru@yahoo.com>
Cc: Tom Lendacky <thomas.lendacky@amd.com>
Cc: David Woodhouse <dwmw2@infradead.org>
Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>