x86/topology: Maintain the total number of sibling in each thread info
topology_sibling_cpumask provides the sibling map of all threads that
belong to one core. When one of its siblings are offlined, the thread count
drops, so that cpuinfo_x86.smt_active becomes false.

Infield Scan technologies (IFS) require to start the test from both threads
of the core before the SCAN can be initiated. Maintaining this count of
number of booted threads of a core can be used in future to detect early
before a test is initiated.

num_threads_active doesn't change due to CPU offline and doesn't
dynamically morph as cpuinfo_x86.smt_active.

Record at boot time the number of active threads and use it to compare when
it matters. For e.g. before a microcode update.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>