x86/microcode: Taint only if late microcode loading is successful
Right now, late microcode loading can, should, and does taint the
kernel. But, that taint will occur even if a late load is rejected
for being unsafe.  A rejected late load is not problematic and
should not result in tainting.

Taint only when the loading of microcode was successful.

Suggested-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Ashok Raj <ashok.raj@intel.com>
---
tglx: Removed goto's and added helper to make code easy to read.