x86/microcode: Do not call apply_microcode() on sibling threads
Microcode updates are applied at the core, so an update to one HT
sibling is effective on all HT siblings of the same core.
apply_microcode() is aware of this and will not attempt to apply
microcode on a sibling where the "lead" thread has already performed
the update.

But in the case when the "lead" thread fails to do the update, the
secondary's revision is not updated.  As it stands, the secondary will
attempt to apply the patch even though the primary is not in a
rendezvous loop.  This is problematic, because the whole rendezvous
dance is to prevent updates when one of the siblings could be
executing arbitrary code.

Never attempt to apply_microcode() on a sibling.  Instead, just
update the CPU version cache in case an update did occur.

Suggested-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Ashok Raj <ashok.raj@intel.com>
Signed-off-by: Dave Hansen <dave.hansen@linux.intel.com>
Cc: LKML <linux-kernel@vger.kernel.org>
Cc: x86 <x86@kernel.org>
Cc: Ingo Molnar <mingo@kernel.org>
Cc: Tony Luck <tony.luck@intel.com>
Cc: Dave Hansen <dave.hansen@intel.com>
Cc: Alison Schofield <alison.schofield@intel.com>
Cc: Reinette Chatre <reinette.chatre@intel.com>
Cc: Thomas Gleixner (Intel) <tglx@linutronix.de>
Cc: Tom Lendacky <thomas.lendacky@amd.com>
Cc: Stefan Talpalaru <stefantalpalaru@yahoo.com>
Cc: David Woodhouse <dwmw2@infradead.org>
Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Cc: Jonathan Corbet <corbet@lwn.net>
Cc: Rafael J. Wysocki <rafael@kernel.org>
Cc: Peter Zilstra (Intel) <peterz@infradead.org>
Cc: Andy Lutomirski <luto@kernel.org>
Cc: Andrew Cooper <Andrew.Cooper3@citrix.com>
Cc: Boris Ostrovsky <boris.ostrovsky@oracle.com>
Cc: Martin Pohlack <mpohlack@amazon.de>