x86/microcode: Mark the primary thread as lead_thread
SMT threads share a CPU core, that includes the microcode. It is only
required to apply microcode to one of the threads of a core. The other
sibling will automatically reflect the new updated revision.

__reload_late() is run on each thread in the system, but only the primary
thread updates the microcode, and sibling simply waits for the update to
complete.

Currently the primary thread is selected via

       cpumask_first(topology_sibling_cpumask(cpu)) == cpu

and the sibling thread via

       cpumask_first(topology_sibling_cpumask(cpu)) != cpu

While this is perfectly functional, one of the CPUs will fall
into one of these blocks. But this code is hard to follow and understand.

To make this perfectly clear, mark the primary thread as 'lead_thread'.
This also avoids calculating the cpumask a second time to filter the
sibling.

No functional change.

Signed-off-by: Dave Hansen <dave.hansen@linux.intel.com>