x86/microcode/amd: Consolidate revision queries
The microcode revision comes out of half of an MSR.  The other half
is unused, but our MSR API requires passing variables for both
halves.  That results in unused 'dummy' variables at each RDMSR site.

Add a wrapper for fetching the microcode revision and use it instead
of RDMSR directly.  This eliminates all but one of the dummy variables.

Signed-off-by: Dave Hansen <dave.hansen@linux.intel.com>