x86/microcode/intel: Always get the current revision from the CPU
collect_cpu_info() returns the current cached values for the revision.
Currently the cached values are updated in when apply_microcode_intel().

When the primary thread updates the microcode, the revision of the sibling
automatically updates since the microcode is shared between the two threads
of the core. But currently its required to send the sibling to
apply_microcode() just to get the revision updated.

Functionally the code does the correct thing, i.e to update the current
values from apply_microcode(). But it appears a little odd, since we do all
the global synchronization before we update to keep the siblings silent
while the primary thread completes the update.
On the other hand when we send the sibling to apply_microcode(), the
primary is already released to the wild.

Always fetch the current revision when collect_cpu_info() is called.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>
---
TBD: Can we nuke the platform id after we fetch it once? That is never
     going to change.