x86/microcode: Introduce new ops get_current_rev()
Introduce a new microcode ops get_current_rev(). This will always return
the real revision from the CPU.

Revision is stored in three places in the kernel.

#1 struct ucode_cpu_info which is used mostly internally to the microcode
   subsystem.

#2 struct cpuinfo_x86 which is used in several other places in the kernel.

#3 Some functions use struct cpu_signature to obtain the revision via
   ops->collect_cpu_info() which simply returns what is already in
   cpuinfo_x86.

Goal is to completely eliminate collect_cpu_info() since its information
otherwise directly available in a system wide cpuinfo_x86.

In subsequent patches, the vendor specific code will not refresh the
cpuinfo_x86 and only updated in common code in core.c

This will also help to not send the sibling to ops->apply_microcode()
in __reload_late(), only to update the current revision. Since the primary
thread of the core that updated the microcode will automatically reflect
the new revision in the sibling.

Add a new ops to only get the current revision.

Suggested-by: Dave Hansen <dave.hansen@intel.com>
Signed-off-by: Ashok Raj <ashok.raj@intel.com>