x86/microcode/intel: Return UCODE_ERROR instead of UCODE_NFOUND when data is inconsistent
Currently when processing microcode files, the return codes are generic and
always return UCODE_NFOUND when the failure is due to some inconsistencies
in the data.

Instead of always returning UCODE_NFOUND when there is an error, return
UCODE_ERROR instead.

Follow up changes are required to make the error code returned to user
space distinguishable.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>
---
tglx: More code simplification