x86/microcode/intel: Fix proper error codes for user space consumption
Currently when processing microcode files, the return codes are generic and
always return -EINVAL to user when the file isn't there. Its difficult to
report a missing file vs a bad file.

Instead of always returning UCODE_OK when a file is not found, return
UCODE_ERROR. Subsequent patch will set proper error codes to be returned to
user space.

Suggested-by: Boris Petkov <bpetkov@alien8.de>
Signed-off-by: Ashok Raj <ashok.raj@intel.com>