x86/reboot: Kick offline CPUs out of mwait before kexec
Offline CPUs are placed in mwait for best power savings. During kexec
reboot, the REBOOT_VECTOR IPI from stop_other_cpus() is only sent to online
CPUs, which are subsequently parked in native_hlt().

CPUs in offline state continue in mwait. The REBOOT_VECTOR IPI also
doesn't make it to offline CPUs since interrupts are disabled.

Interrupts and cache flush can cause the CPU to drop out of mwait.
After the new kernel is loaded the previously offlined CPUs are still
using text and data from the previous kernel image. Once the new kernel is
launched, page-tables, data are overwritten by the new kernel.

This causes the CPU's in mwait to drop out of mwait and proceed to the next
instruction. But alas, even the page-tables are gone by now. When the CPU
is attempting to fetch code, that results in a fault, the fault handler is
also gone ends up with a double and then a triple fault.

The core signals a MCE and enters shutdown state. The core # is recorded
in a globally visible MCE bank in that socket for reporting purposes.

When system is in nosmt, since every core is reporting an error. This
causes multiple machine checks causing reset conditions even before any
information can be collected.

When only one thread is offline, it was simple to track down the core
reporting MCE was indeed the one that had an offline thread.

Moving the offline thread for each experiment confirmed that the reporting
core moved along with it.

Registers during MCE.

eax=00000020    ebx=00000000    ecx=00000000    edx=00000000
esi=90a9aa46    edi=90aa0704    ebp=88920000    esp=48717ec0
eip=8f066796
^^^^^^^^^^^^ Faulting ip.

cs=0010    ds=0000    es=0000    fs=0000    gs=0000
eflags=00010046

>>>  ##Dumping Thread stack for 20 instructions, before and after
0x10:ffffffff8f066782    85c0             test eax, eax
0x10:ffffffff8f066784    7e07             jle $+0x09  ;a=0xffffffff8f06678d
0x10:ffffffff8f066786    0f002df3685a01   verw word ptr [rip + 0x015a68f3]
0x10:ffffffff8f06678d    4c89e0           mov rax, r12
0x10:ffffffff8f066790    4889d9           mov rcx, rbx
0x10:ffffffff8f066793    0f01c9           mwait <--- mwait_play_dead::mwait
0x10:ffffffff8f066796    e8e5980701       call $+0x010798ea  ;a=0xffffffff900e0080
-------------^^^^^^^^^^^^^^^^^^^^^^^^^^^ The call to cond_wakeup_cpu0

0x10:ffffffff8f06679b    85c0             test eax, eax
0x10:ffffffff8f06679d    75c6             jnz $-0x38  ;a=0xffffffff8f066765

Some Hitchcock work from Thomas :-) below that root causes with more
evidence.

Analysis:

Based on the analysis below, after the first kexec from a cold/warm
boot seems to survive since the new kernel is loaded at a different
address. Subsequent kexec seems to overwrite the previous kernel text
and also overwrites the monitored address, causing the mwait to exit.
This leads to triple faults and mayhem breaks down.

In order to avoid this scenario, release all offline CPUs from mwait and
park them in native_halt(). This is relatively safe until some NMI/MCE/MCE
fires, which can trigger similar scenario.

Analysis:

First thing I looked at was the physical address at which
mwait_play_dead() resides (KASLR disabled on the command line):

Cold start:     0x00010f6f80
1st kexec:      0x303ccf6f80
2nd kexec:      0x303ccf6f80

So kexec() on a kernel which was kexec()'ed already ends up overwriting
the previous kernel text. That's obviously a horrorshow when mwait drops
out unless it's exactly the same kernel image and KASLR is disabled and
the page tables of the previous kernels are not overwritten. Lot's of
conditions to be true...

Now I added a printk to dump the physical address of the cache line
which is monitored by mwait_play_dead(), which is thread info::flags of
the idle task of the offlined CPU. Further I added a 'deadmagic' field
into thread_info which is set to 0xDEADBEAF by the offlined CPU before
entering the mwait loop. I validated and printed that on the control CPU
after the outgoing CPU reported death.

Then I added a command line option to the kernel, which takes the
physical address of the monitored cache line as argument. That makes the
boot CPU of the kexec kernel dump the deadmagic field.

Offline CPU:	0xdeadbeaf
kexec dump:     0x00000000

So the kexec kernel overwrites the monitored memory, which obviously
causes the offline CPU to drop out of mwait.

Now as long as neither the kernel text nor the relevant page table of
the offline CPU have been overwritten this works by some definition of
works.

If either of these was overwritten, then the offline CPU goes into lala
land, which obviously can end up in a machine check or insta triple
fault. I did some experiments to that effect and while I wasn't able to
get a machine check, I managed triple faulting reliably.

Also when keeping all CPUs online I validated that the kexec kernel
overwrites the page table of the previous kernel, which is a good enough
explanation why this triple faults after coming out of mwait...

The reason why this always happens at the same point during boot of the
kexec kernel is that it is always the same part of bringup which
happens to overwrite something in the previous kernel.

Now let's look at how kexec() works:

kernel_kexec()
  if (CONFIG_KEXEC_JUMP) {

     // Not taken

  } else {
     ...
     migrate_to_reboot_cpu();
     ...
     machine_shutdown()
      stop_other_cpus()
  }

  machine_kexec()
     ....
     relocate_kernel()  <- Jumps to the new kernel

So stop_other_cpus() kicks the non-boot CPUs into stop_this_cpu() which
ends up with

     for (;;)
         native_halt();

Which is supposedly safe under the assumption that there is no MCE, NMI,
SMI which kicks the stopped CPU out of HLT...

Obviously all offline CPUs are not brought out of their respective
play_dead() which results in the above if they are playing dead with
mwait.

So that's a plain bug in x86/kexec() forever, but even with HLT it only
works by some definition of works...

I wasn't able to reproduce in a VM, but on my SKL-X with nosmt the first
kexec() survives, but the second one where the kernel is overwritten
insta triple faults.

Q: Why can't we have some simple and reliable STOPDEAD mechanism?
A: Because this can be handled in software.

Sure we can. Just that it requires to squint enough so you can claim
it's reliable...

Fixes: 76fac077db6b ("x86, kexec: Make sure to stop all CPUs before exiting the kernel")
Reported-by: Ashok Raj <ashok.raj@intel.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Ashok Raj <ashok.raj@intel.com>