x86/microcode: Keep offline threads in busy wait until microcode update completes
Kernel parks the offline threads in mwait for best possible power savings.
When updating microcode those sibling threads must not be executing any
complex instruction, that includes mwait. Since mwait is a microcoded
instruction and could be patched as a result of the microcode update.

Briefly park the offline threads in busy wait until the microcode loading
has completed and then place them back in mwait.

Suggested-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Ashok Raj <ashok.raj@intel.com>