x86/microcode: Add a flag to force load on the secondary thread
When microcode is updated on the primary thread of a core, the secondary
thread also receives the update and the hardware updates the revision
number appropriately. But AMD has certain states that is fixed up only
when the sibling thread is also updated. This requires the update to be run
sibling CPUs even if the revision number is already reflecting the correct
value.

Add a control flag in `microcode_ops` to let common code perform an update
for the secondary thread.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>