x86/microcode/intel: Add enumeration support for microcode uniform update
New Intel processors have a new capability that simplifies microcode
loading and also reduces a lot of the race conditions that exists in
current microcode loading.

When Uniform Loading is enumerated and configured, loading the microcode in
one thread or a core will ensure its applied to all cores in the system,
even across sockets.

Add support to enumerate the capability and some basic definitions.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>
Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>