x86/microcode: Apply microcode on siblings only when update is successful in primary CPU
After the primary CPU is updated, the sibling CPU is also updated in
__reload_late(). This is required to update the revision ID in cpuinfo_x86,
so /proc/cpuinfo will display the newly updated microcode revision for both
the primary and secondary CPUs of the core. But it isn't necessary if the
update on primary failed.

Apply microcode for sibling only when primary CPU update was successful.

Suggested-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Ashok Raj <ashok.raj@intel.com>