x86/microcode: Handle "nosmt" correctly
CPUs which are parked due to software SMT disable sit in a MWAIT loop,
which is not suited for late loading. Obviously they cannot run the stopper
thread either. Prevent to late load in this case.

This might be cured by onlining them up to CPUHP_AP_SMPBOOT_THREADS, but
that needs some thought.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>